Capstone_COVID19/website/requirements.txt
flask
azure-storage-blob
urllib3
requests
gunicorn